vpn: don't choose route priority for VPN connections based on the parent device
The parent device is just the device that happens to be the best device.
It does not mean, that its route metric should be inherited to the VPN
connection.

This also makes the resulting route metric for VPN connections much
more predictable: now it is either ipv4.route-metric from the connection,
or it falls back to NM_VPN_ROUTE_METRIC_DEFAULT (10).