cli: request an activation without a device if we can't do a good guess
If the user didin't specify an interface and we can't do a good guess just
proceed the activation without the device. Let the server decide which
device to use.

Also, don't even try to guess the device if the connection is already
active -- the server will likely do a better guess.