platform: re-enable the platform link test
The link test was disabled in commit 67ad3fcb5b65fd74cab7f16531fce7ab457d8c43.

The previous issues are not fixed, but apparently disabling the test doesn't
help to get it fixed.

Re-enable it and if it fails we have a better reason to fix it.
Or maybe it works now (?). Didn't fail for me...