dns: fall back to writing resolv.conf when other methods are not available
After commits:

de0d623680d7 dns: don't fall back to other methods when resolvconf or netconfig fail
e573977b8084 dns: allow runtime selection of resolv.conf manager

the method used to write /etc/resolv.conf is specified by a
configuration parameter and NM doesn't try other methods if the chosen
one has not success.

Restore the fallback to the 'none' method if resolvconf or netconfig
executables are not available, so that when the value of 'rc-manager'
is misconfigured the system continues to work properly.

https://bugzilla.gnome.org/show_bug.cgi?id=749286