libnm/vlan: default to vlan.flags=REORDER_HDR for new connection (rh #1250225)
The kernel defaults REORDER_HDR to 1 when creating a new VLAN, but
NetworkManager's VLAN flags property defaults to 0 and thus REORDER_HDR is not
set by default for default NM created VLANs.

We want to match kernel defaults (only for newly created connection). So we
set REORDER_HDR if the flag is not present in D-Bus hash.

https://bugzilla.redhat.com/show_bug.cgi?id=1250225