device: perform duplicate address detection for static addresses (rh #1259063)
When static addresses are to be configured, DAD is done to check whether they
do not clash with IPs already present in the local network. The detection is
performed by ARP using "arping -D". If a response is received for the ARP
reguest, the corresponding IP is duplicated. If no response is received within
a timeout, the IP can be used.

send_arps() was renamed to send_arps_announce() and only announces configured
(not duplicated) addresses.
send_arps_dad() is used for DAD.

DAD is configurable via ipv4.dad-timeout property.

https://bugzilla.redhat.com/show_bug.cgi?id=1259063